[frontend-5] deleted areaInit, added useEffect in AreasSearch
