+++ frontend/src/components/CompanySearch/CompanySearch.js
@@ -11,7 +11,7 @@ import {COMPANY_SEARCH} from '../MainComponent';
 import './CompanySearch.css';
 
 
-const CompanySearch = ({ currentTab, openCompetitorsList, companyId, areaId, plainAreas }) => {
+const CompanySearch = ({ currentTab, openCompetitorsList, companyId, areaId }) => {
 
     if (currentTab !== COMPANY_SEARCH) {
         return null;
@@ -26,7 +26,7 @@ const CompanySearch = ({ currentTab, openCompetitorsList, companyId, areaId, pla
                     <Search/>
                 </div>
                 <div className="search__btn">
-                    <Button onClick={openCompetitorsList} disabled={!companyId && !areaId}>Показать конкурентов</Button>
+                    <Button onClick={openCompetitorsList} disabled={!companyId }>Показать конкурентов</Button>
                 </div>
             </div>
             <div className="history">
@@ -39,7 +39,6 @@ const CompanySearch = ({ currentTab, openCompetitorsList, companyId, areaId, pla
 export default connect(
     state => ({
         companyId: state.search.companyId,
-        areaId: state.search.areaId,
-        plainAreas: state.areaInit.plainAreas
+        //areaId: state.areaSearch.areaId
     }),
 )(CompanySearch);
